Different Singleton implementation
Chose a different implementation of Singleton pattern.
In principle, this pattern should result in a (little bit) faster app
startup, as instances are created only on access. However, as all
(relevant) classes are also created by the Application class on
creation, there probably is no difference. But as this approach does not
need Holder classes, it is still preferred.
Also with this approach we have control over the instantiation order (by
the order in which we call the getInstance() methods for the first
time). The old approach created the instances in a (more or less) random
order.